Update README.md
Research problem
Effective categorization of student involvement using deep learning methods on visual data is the main issue this work aims to solve. This study is motivated by the fact that, often, conventional teaching approaches fall short of meeting the unique levels of student involvement, which may result in less-than-ideal learning results. Before we formally address our study question, it is important to recognize that student involvement has many aspects, including emotional, cognitive, and behavioral aspects that each appear differently in visual representations. 
The choice of this topic stems from the two requirements of improving the efficacy of education and using state-of-the-art AI capabilities to create a more interesting learning environment. The nonverbal clues indicating the minute intricacies of student involvement call for a very granular and flexible analytical method. This work tries to close the gap between the promise of AI-driven analytical tools and the actual educational demands by using sophisticated deep learning architectures. For example, this tool can be used in an online meeting platform like zoom meet or google meet, providing seamless and smart solution to handle overall sentiment of students.